[HW-06] CoinFragment ViewModel 수정
